Add max retries setting
Previous behaviour: If an error occurs while downloading, the program crashes (due to no error handling)

New behaviour: If an error occurs while downloading, the program will retry the download, up to an amount specified in the settings. If it still can't download the file it will gracefully stop the download and notify the user.